backend/requirements.txt
django>=5.2
djangorestframework>=3.14.0
python-socketio[asyncio]==5.11.1
aiohttp>=3.9.3
django-cors-headers>=4.3.1
djangorestframework-simplejwt>=5.3.1
uvicorn[standard]>=0.27.1